+++ cedar-policy-validator/src/lib.rs
@@ -129,18 +129,18 @@ impl Validator {
             None
         } else {
             Some(
-                self.validate_entity_types(p)
-                    .chain(self.validate_action_ids(p))
-                    // We could usefully update this pass to apply to partial
-                    // schema if it only failed when there is a known action
-                    // applied to known principal/resource entity types that are
-                    // not in its `appliesTo`.
-                    .chain(self.validate_action_application(
-                        p.principal_constraint(),
-                        p.action_constraint(),
-                        p.resource_constraint(),
-                    ))
-                    .map(move |note| ValidationError::with_policy_id(p.id().clone(), None, note)),
+                // We could usefully update this pass to apply to partial
+                // schema if it only failed when there is a known action
+                // applied to known principal/resource entity types that are
+                // not in its `appliesTo`.
+                self.validate_action_application(
+                    p.principal_constraint(),
+                    p.action_constraint(),
+                    p.resource_constraint(),
+                )
+                .map(move |note| ValidationError::with_policy_id(p.id().clone(), None, note))
+                .chain(self.validate_entity_types(p))
+                .chain(self.validate_action_ids(p)),
             )
         }
         .into_iter()
@@ -301,10 +301,17 @@ mod test {
                 Some("Action::\"action\"".to_string()),
             ),
         );
+
         assert!(!result.validation_passed());
-        assert!(result.validation_errors().any(|x| x == &principal_err));
-        assert!(result.validation_errors().any(|x| x == &resource_err));
-        assert!(result.validation_errors().any(|x| x == &action_err));
+        assert!(result
+            .validation_errors()
+            .any(|x| x.error_kind() == principal_err.error_kind()));
+        assert!(result
+            .validation_errors()
+            .any(|x| x.error_kind() == resource_err.error_kind()));
+        assert!(result
+            .validation_errors()
+            .any(|x| x.error_kind() == action_err.error_kind()));
 
         Ok(())
     }
@@ -379,6 +386,7 @@ mod test {
             ast::EntityUID::from_components(
                 "some_namespace::Photo".parse().unwrap(),
                 ast::Eid::new("foo"),
+                None,
             ),
         );
         set.link(
@@ -397,6 +405,7 @@ mod test {
             ast::EntityUID::from_components(
                 "some_namespace::Undefined".parse().unwrap(),
                 ast::Eid::new("foo"),
+                None,
             ),
         );
         set.link(
@@ -432,6 +441,7 @@ mod test {
             ast::EntityUID::from_components(
                 "some_namespace::User".parse().unwrap(),
                 ast::Eid::new("foo"),
+                None,
             ),
         );
         set.link(
@@ -450,7 +460,9 @@ mod test {
             None,
             ValidationErrorKind::invalid_action_application(false, false),
         );
-        assert!(result.validation_errors().any(|x| x == &invalid_action_err));
+        assert!(result
+            .validation_errors()
+            .any(|x| x.error_kind() == invalid_action_err.error_kind()));
 
         Ok(())
     }